fix bucket.Create method description (#201)
* fix bucket.Create method

Escrow `Bucket.Create` method and documentation is updated to be acurate.

fix #170

* replace escrow.Bucket.Create with Build method

Rename `Create` to `Build` and update documentation.

Create method should persist returned object in the store. Because of
how escrow extension is implemented, a zero value escrow instance cannot
be persisted due to validation.
Escrow controller is expecting a zero amount escrow instance, that it
will deposit initial amount to. Rename `Create` to `Build` so that the
logic can remain as is, but the confusing naming is fixed.