new error pattern proposal
As part of this proposal, `x/paychan` extension was updated to use a new
error reporting pattern. This gives a good understanding of the change,
but does not show its full potential.

This change is implementing all requirements described during the last
meeting
https://groups.google.com/a/iov.one/d/msg/blockchain-dev/xQqsZml9YhM/mew5nMhIDQAJ

There is no `x/<extension>/errors.go` file anymore. No error is declared
upfront. Instead `werrors.E` function is used everywhere. `werrors` is a
work-in-progress package name.

Code declarations are limited to `werrors` package. Extensions no longer
declare their own, specific error codes.
My motivation for this change was that too granular error codes require
too much work to handle. For example, instead of having an error code
for the `paychan.CreatePaymentChannelMsg.timeout` field being less than
zero, use `InvalidMsg` code and provide text description `"timeout must
be greater than zero"`.

I think that bubbling up internal errors and exposing them via public
API is a bad idea. For example, if a database could not perform an
operation, client must never receive message `"leveldb/storage: already
locked"`. This is an implementation details that must not be leaked out
to the client.
Above design is propagated throughout the codebase already and removing
it is out of scope of this proposal. This change is only attempting to
minimize the exposure of it.
Above issue is addressed by splitting error codes into public and
private ranges.  Everything between 0 and 1000 is a private error for
which `ABCILog` method returns a generic information. Error codes of
1001 and higher are passing to the client the exact error content.

`werrors.E` was modeled after upspins `errors.E`
(https://godoc.org/upspin.io/errors#E). I is using `...interface{}` as
an argument to provide an easy to use and extensible API. Adding more
functionality is very simple.
Drawback of using such generic interface is that compiler cannot
validate parameters and some mistakes can be captured only during the
runtime.
The same problem of `interface{}` arguments exist for all `XXXf`
functions of the standard library.  There the solution is `go vet` which
can detect misuse and report the problem.

This implementation does not include traces, but adding them requires
changes only to the `werrors` package.