new error pattern proposal
This change build upon commit c16131cfabd8e58c5587ea1efc3fec5f9deb3732.
Please read it first to get the full understanding of the changes made.

This change is implementing all requirements described during the last
meeting
https://groups.google.com/a/iov.one/d/msg/blockchain-dev/xQqsZml9YhM/mew5nMhIDQAJ

Function `werrors.E(...interface{}) error` was replaced with two new
functions that have a notation that allows the compiler to validate the
correctness of their use.

`werrors.New` is taking an error and a description. It is creating a new
error instance. This function is to be used only when an error situation
is not caused by an another error or when we want to ignore previous
error information.

`werrors.Wrap` requires an `error` instance and a description. It is
creating an error instance that wraps given one with more information.
Having description of each frame operation results in more detailed
error messages.
Notice that passed `error` can be both an error created with
`werrors.New` in which case it contains an error code or any other
implementation, that does not provide error code.
If an error code can be deducted, it is reported by wrapper as well.
Otherwise an error code of value 0 is used (Internal error).

Above design has an interesting property. We do not want to expose
internal errors. So to expose an error we must explicitly specify its
error code to be greater than 1000.
Errors coming from libraries are always internal, because they do not
provide `ABCICode` method and their code will default to 0 (Internal
error).