Initial global configuration package implementation
`gconf` package implements a set of operations on a global
configuration.

Configuration is stored in a database with an interface compatible with
`weave.KVStore`. Each value stored by configration is serialized using
JSON.

To load a configuration value, use one of the provided functions (`Int`,
`String`, etc.). In case of parsing issues they log an error and return
empty value for given type. This should not happen, unless database was
manually altered.

Next step is to add a support for runtime configuration update via
message sending.

resolve #202