new error pattern proposal (2) (#275)
* new error pattern proposal

As part of this proposal, `x/paychan` extension was updated to use a new
error reporting pattern. This gives a good understanding of the change,
but does not show its full potential.

This change is implementing all requirements described during the last
meeting
https://groups.google.com/a/iov.one/d/msg/blockchain-dev/xQqsZml9YhM/mew5nMhIDQAJ

There is no `x/<extension>/errors.go` file anymore. No error is declared
upfront. Instead `werrors.E` function is used everywhere. `werrors` is a
work-in-progress package name.

Code declarations are limited to `werrors` package. Extensions no longer
declare their own, specific error codes.
My motivation for this change was that too granular error codes require
too much work to handle. For example, instead of having an error code
for the `paychan.CreatePaymentChannelMsg.timeout` field being less than
zero, use `InvalidMsg` code and provide text description `"timeout must
be greater than zero"`.

I think that bubbling up internal errors and exposing them via public
API is a bad idea. For example, if a database could not perform an
operation, client must never receive message `"leveldb/storage: already
locked"`. This is an implementation details that must not be leaked out
to the client.
Above design is propagated throughout the codebase already and removing
it is out of scope of this proposal. This change is only attempting to
minimize the exposure of it.
Above issue is addressed by splitting error codes into public and
private ranges.  Everything between 0 and 1000 is a private error for
which `ABCILog` method returns a generic information. Error codes of
1001 and higher are passing to the client the exact error content.

`werrors.E` was modeled after upspins `errors.E`
(https://godoc.org/upspin.io/errors#E). I is using `...interface{}` as
an argument to provide an easy to use and extensible API. Adding more
functionality is very simple.
Drawback of using such generic interface is that compiler cannot
validate parameters and some mistakes can be captured only during the
runtime.
The same problem of `interface{}` arguments exist for all `XXXf`
functions of the standard library.  There the solution is `go vet` which
can detect misuse and report the problem.

This implementation does not include traces, but adding them requires
changes only to the `werrors` package.

* new error pattern proposal

This change build upon commit c16131cfabd8e58c5587ea1efc3fec5f9deb3732.
Please read it first to get the full understanding of the changes made.

This change is implementing all requirements described during the last
meeting
https://groups.google.com/a/iov.one/d/msg/blockchain-dev/xQqsZml9YhM/mew5nMhIDQAJ

Function `werrors.E(...interface{}) error` was replaced with two new
functions that have a notation that allows the compiler to validate the
correctness of their use.

`werrors.New` is taking an error and a description. It is creating a new
error instance. This function is to be used only when an error situation
is not caused by an another error or when we want to ignore previous
error information.

`werrors.Wrap` requires an `error` instance and a description. It is
creating an error instance that wraps given one with more information.
Having description of each frame operation results in more detailed
error messages.
Notice that passed `error` can be both an error created with
`werrors.New` in which case it contains an error code or any other
implementation, that does not provide error code.
If an error code can be deducted, it is reported by wrapper as well.
Otherwise an error code of value 0 is used (Internal error).

Above design has an interesting property. We do not want to expose
internal errors. So to expose an error we must explicitly specify its
error code to be greater than 1000.
Errors coming from libraries are always internal, because they do not
provide `ABCICode` method and their code will default to 0 (Internal
error).

* update errors proposal

- only zero Code represents an internal error
- tests

* Update werrors package.

- remove Code.String method
- add tests

* Upgrade werrors package.

This is an upgrade an quite some change to the previous proposal of
errors refactoring.

If this change is accepted, `werrors` package content will be moved to
`errors`.

* Move werrors code to errors package.

Original weave `errors` package provides `Wrap` and `New` functions with
notations not compatible with `werrors` API.
As a transision solution versions of those two functions with more
permissive notation was created. It does compile with both old and new
notation beings used. Once the transition is complete, it must be
removed.

At this stage, new errors implementation does not provide stack trace
implementation. Original code was copying and importing `pkg/errors`
package for this. I think the same can be achieved without importing
`pkg/errors` and removing this external package from weaves hard
dependencies completely.

* Deprecate `errors.New`.

Remove `New` compability layer as this function is not needed in the new
errors API.

* rename error components as requested

* remove panic from temporary function

* errors.Wrap(nil, "") returns nil

This is update to bring back nil wrapping as it used to work in the old
implementation.
This change is required for
https://github.com/iov-one/weave/pull/305#discussion_r255043779

* Feature/proper naming and legacy (#308)

* refactor registerNew -> register

* refactor legacy stuff, tests still failing

* fix names

* Wrap returns nil on nil input, comment out stacktrace tests

* Fix up debug log output

* Don't return 'internal error', fix tests

* Special code for errors from panic recovery

* Redact panic errors at abci interface

* Happify go-lint