relace "USERNAME" with username.ModelName
Declare a model name for username nft implementation. This provides more
clarity of what that string value is.

Although it used to be all uppercase due to being a protobuf enum
declaration, there is no need to keep it that way. Because this is not
production used functionality this value can be made pretty.