remove gconf calls cache
Caching configuration in a decorator makes it immutable and prevents
from being updated during the runtime.

If the performance is a concern, instead of caching in the decorator a
store wrapper could be used that would keep an in memory representation
of the whole gconf set and flush on any update.