wip: refactoring weave test helpers
A new package `weavetest` is created. It contains all the commonly used
functions and structures for tests.
All one time used mocks were moved next to the test using them.

This is a work in progress state. No documentation and not all
functionality was ported/refactored.

resolve #372 and much more! :muscle: