migration: initial version
This is a initial version and a concept description rather than a
functional code of a schema migration package.

Each package that must register a version migration for each model and
mesage owned. Migration can be a no operation migration. For example

  // Enable message
  migration.MustRegister(0, &MyMsg{}, migration.NoModyfication)
  // Enable model
  migration.MustRegister(0, &MyModel{}, migration.NoModyfication)

  migration.MustRegister(1, &MyMsg{}, migration.NoModyfication)
  migration.MustRegister(1, &MyModel{}, func(weave.Context, weave.KVStore, migration.Payload) { ... })

  migration.MustRegister(2, &MyMsg{}, func(weave.Context, weave.KVStore, migration.Payload) { ... })
  migration.MustRegister(2, &MyModel{}, func(weave.Context, weave.KVStore, migration.Payload) { ... })

Migration version are tracked per package so each time a single (or
multiple) message schema is updated, a schema migration function must be
declared for each message and model type. This choice is made as opposed
to allowing a missing migration in order to avoid silent errors when a
migration succedded due to a missing migration function and application
fails much later due to inconsistent data.

To migrate a message or a model to desired version, `Migrate` it

  mymsg := &MyMsg{...}
  err := migration.Apply(ctx, db, mymsg, desidedVersionNumber)

Apply function is supposed to be embeded in the weave handler and bucket
implementation to avoid manual version upgrades.