x/cash, x/escrow x/distribution: use schema versioning
This is a big change that breaks the extensions. Add integration with
the migration package to allow schema versioning for models and
messages.

This is clubbed all together because all those packages tightly require
x/cash to provide their functionality.

Small adjustments to the migration package to make error messages
better.

part of #564