migration: introduce a handler router based on message schema version
A new handler implementation is introduced.
`SchemaRoutingHandler` is a container that clubs toghether message
handlers for a single type message but different schema formats. Each
handler is registered together with the lowest schema version that it
supports. For example:

    handler := SchemaRoutingHandler{
      1: &MyHandlerVersionAlpha{},
      7: &MyHandlerVersionBeta{},
    }

In the above setup, messages with schema version 1 to 6 will be handled
by the alpha handler. Messages with schema version 7 and above are
passed to the beta handler.

It is not allowed to use an empty SchemaRoutingHandler instance. It is
not allowed to register a handler for schema version zero.
All messages processed by this handler must implement Migratable
interface.