x/paychan: use wall clock for the timeout
Use the wall clock value to determine a Payment Channel timeout in the
`x/paychan` extension. This is a change similar to the one done in
`x/escrow`.

`escrow.isExpired` was copied to the root level `weave` package as
`IsExpired`. Next change is to update `x/escrow` to use it instead of
the local implementation.

resolve #481