orm: proposal of ModelBucket implementation
This is a proposal and a proof-of-concept implementation of a bucket
that operates directly on the model rather than Object instances.

This implementation relies on the Bucket to access the database. Final
implementation must not use Bucket but directly use KVStore instance.

Methods used (`One`, Put`) are to be refined. I want to use something
very different than `Get` and `Save` of the existing bucket
implementation to avoid confusion and conflicts.

There are some more topics that are to be addressed before this
implementation is finished:

- [ ] integration with the `migration` package
- [ ] do not use Bucket but directly access KVStore
- [ ] register for queries
- [ ] implement `Many` method for getting a list of models
- [ ] implement `Has` method for checking if an entity exist without
      deserialization
- [ ] integrate in at least one extension to learn if the new interface
      looks good and is easy to use

resolve #536