orm: update NewModelBucket API
When creating a new model bucket using `NewModelBucket`, instead of
explicitely requiring an `orm.Bucket` to be passed, accept bucket name,
model reference and optional parameters. `orm.Bucket` instance is
created transparently for the user and never exposed.

This change allows to directly refence represented model and provide
runtime type checks required for some of the methods. It is also
providing a simpler and extensible API that is no longer dependant on
the `orm.Bucket` implementation.

This is addressing
https://github.com/iov-one/weave/pull/615#discussion_r283962361