migration: provide schema migrating registry decorator
It is common to decorate with SchemaMigratingHandler all handlers before
registering. This requires quite some typing and passing everywhere the
same package argument. To simplify this proces a new decorator is
introduced: `SchemaMigratingRegistry`.

The old way of registering a schema migration for a handler is:

	var r weave.Registry
	r.Handle("a", SchemaMigratingHandler("mypkg", myHandlerA)
	r.Handle("b", SchemaMigratingHandler("mypkg", myHandlerB)

The above is equivalent to the below code enabled by this change:

	r = SchemaMigratingRegistry("mypkg", r)
	r.Handle("a", myHandlerA)
	r.Handle("b", myHandlerB)