orm: update ModelBucket.ByIndex to return result keys
In addition to unpacking data into given slice, return the list of keys
in the same order as the corresponding value unpacked to the slice.

Model does not contain the key. Matching it by an index with returned
list of keys is the only way to determine each model primary key value.

This is addressing comment
https://github.com/iov-one/weave/pull/615#discussion_r283956288