cmd/bnsd/scenarios: redo test to not use global state
This is a huge infrustructural change. No test functionality was
modified.

Instead of using the `TestMain` to setup test runner once and reuse the
state between the test cases, provide `bnsdtest.StartBnsd` function.

`TestMain` and any global state was removed. Each test case must
explicitely call `bnsdtest.StartBnsd` to create an independent bnsd
instance and use it for the test.

It is still possible to connect to a remote bnsd node for running tests,
but because of a setup requirement it might no longer make sense for
some test cases.

This change is motivated by the
https://github.com/iov-one/weave/issues/563 issue, where we must use a
non zero minimum fee value while other tests cases require it to be
zero.