cmd/bnsd/scenarios: allow to configure bnsd instance
This change introduces a functional parameters to configure how the bnsd
is started.

Because of clients spaghetti code (everything is a client!) it is very
hard to reference the right client instance in the EnvConf instance.
This change is introducing more complexity by adding `TendermintClient`
method to satisfy some of the interfaces.