Removed switch abuse, where it stood in for a series of if statements
There are some valid switches, where it just takes on value (enum)
and matches, and I left that. Note that these were used heavily
in Validate (to save a bit of typing) but would have to change anyway
for multi-error as many of these could be true and we would want to
match them all.

Sorry @alpe