cmd/bnscli: update sequence parsing
When parsing a user provided sequence value, require explicit format
inforamtion.
Before the `unpackSequence` function was trying to guess the format. The
problem is that for example `1234` is a both valid decimal, hex and
base64 representation of a value. This could lead to false positives
when parsing.

With this change, always default to the decimal parser. Allow to provide
a prefix (ie `hex:` or `base64:`) to provide a value using non default
encoding.