errors: huge refactor of how multi error is implemented
This is a big change to how multi-error is implemented internally with a
minimal change to the API.

In the previous implementation, the `multiErr` type was exposed throughout
the `errors` package. Many places were type casting to check if an error
is a `multiErr` instance and provided as special execution path for this
case. This result in an entangled implementation that is hard to extend.
Adding multi-error was breaking the previous pattern where the error
implementations (`Error`, `wrappedError`, stack trace) were not aware of
each other implementation and were using only common interfaces.

This change is a preparation to implement the `Field` function as
discussed in https://github.com/iov-one/weave/issues/730

This change removes all dependency and type check. A new interface
`unpacker` is added that allows to unpack an error and transform it to a
group of errors it represents.
This design has a few consequences:
- multi-error implementation is independent of any other error
  implementation
- multi-error implementation is not restricted to `weave/errors` package
- errors cannot be flattened as we cannot do the typecasting. This was
  possible in the previous implementation and could cause unwanted
  behavior. Nested errors form a tree structure
- because nested errors form a tree structure, ABCI information can no
  longer return a flat list of errors. They could return a tree
  structure containing this information (not implemented)
- `Redact` function silence the whole group of errors that contain
  `ErrPanic`. It is not possible to hide only the true `ErrPanic`. This
  is logically correct as multi-error containing `ErrPanic` *is**
  `ErrPanic`