Allow to use !nochangelog to mark pull request.
We use `#trivial` to allow merging a pull request without any
modyfications to the CHANGELOG file.

In addition, allow to use `!nochangelog` string to achieve the same
result.
`#trivial` is triggering GitHub's stupid autocompletion. It might also
be confusing to people from outside of the project what it means.

Do not search for any of those strings in the pull request title as we
never put it there.