x/sigs: add benchmark for signature verification (#942)
A new benchmark that tests `sigs.VerifySignature` function. This
function is called for every signed transaction, for each signature that
this a transaction provides. In the current implementation this function
is called separately for every signature before handlers `Check` and
`Deliver` callbacks.

Running on my local machine, results are:

	goos: linux
	goarch: amd64
	pkg: github.com/iov-one/weave/x/sigs
	BenchmarkVerifySignatures/memory_store_small_payload-4                  10000   174463 ns/op
	BenchmarkVerifySignatures/memory_store_big_payload-4                     5000   285872 ns/op
	BenchmarkVerifySignatures/hard_drive_backed_store_with_small_payload-4  10000   173257 ns/op
	BenchmarkVerifySignatures/hard_drive_backed_store_with_bit_payload-4    10000   237035 ns/op

There is no significant difference between in memory store and hard
drive backed one because all operations are cached and flushed only on
commit.

There is a difference between small payload (around 20 bytes) and big
payload (1024 times bigger than small payload).

For a small payload, it takes 0.17ms to verify a signature. It takes
less than 0.3ms to verify a huge payload.

It is hard to predict the time saved by optimizing signature checking
validator to verify each signature only once. This is because this time
is dependant on the computational power of the host and on the number of
signatures a transaction contains.

Assuming that the machine running a weave application is comparable to
my laptop and each transaction contains only one signature the gain
would be around 0.2ms faster transaction processing.

resolve #940