bnsd/username: use gconf to manage username validation rules
Instead of using hardcoded set of validation rules for the username, use
gconf package to manage validation rules dynamically.
Using gconf has another advantage that we can change the validation
rules during the lifetime of a network without a downtime. Because
changing the validation rule is part of the blockchain, no additional
code is required to support this scenario.

resolve #952