x/msgfee: add message to change the message fee value (#992)
* x/msgfee: add message to change the message fee value

`x/msgfee` was extended to provide a message to set a fee for a given
message path. Use zero value amount to delete a fee. Transaction for
setting a message fee must be signed by the fee administrator.
Fee administrator is set in the `x/msgfee` configuration, using gconf
and should be initialized with the genesis file.

The messsage fee setting message was added to `bnsd` and `bnscli`.

* x/msgfee: add message to change the message fee value (#996)

* bnscli: introduce changes to genesis.json to match the one from bnscli

* tmtest: fix cleanup to make sure processes were properly killed before running the next test

* changelog: update