x/lateinit: allow to initialize an entity
x/lateinit provides a way to declare and execute a one time state
modification by creating an entity. This extension is intended to be
used in a similar fashion to genessis, but for already initialized
chains.

Lateinit provides a way to create a defined in code entity in order to
change the current state of the chain. To apply such modification two
steps are required.
1. Declare a state migration in code and build new bnsd binary. Distribute the
   new binary to all validators.
2. Send an `lateinit.ExecuteInitMsg` with information which
   initialization to execute.

Once an initialization is declared in code and executed, it must never
change. This is crucial for the state rebuilding.

Each initialization can be executed only once. Subsequent execution
attempts will fail.