x/blueaccount: a new message to delete all domain accounts
A new message was added `DeleteAllAccountsMsg` to the `x/blueaccount`
extension that allows a domain owner to delete all (except the one with
an empty name) accounts that belong to that domain.

The above message can be used together with a domain transfer message in
order to transfer ownership of a domain without any of the accounts that
might have belonged to it before.

resolve #1036