datamigration: package created
A new package `datamigration` was created.  `datamigration` allows to
create and execute data modification on chain. Each migration is
declared as a Go code, that can access and modify database state.

This functionality is helpful when migrating state of an existing code
to a new implementation or when an initialization of a newly installed
extension is needed, but the genesis cannot be used.

resolve #1048