bnsd/x/preregistration: a new extension for domain preregistration
A new extension `cmd/bnsd/x/preregistration` was implemented and
integrated in `bnsd`. It allows to collect list of (domain name,
address) pairs.
In addition a data migration was registered in `bnsd` to rewrite all
registered domain names from `preregistration` extension into `account`.

Preregistration extension is a one time approach of preregistering
domains, without a need for hardcoding the list of domains we want to
register.
It is the registrator responsibility to ensure that there is no conflict
between preregistration data and already registered account domains.

resolve #1072