no explicit username token registration owner
This change is a reverse of modification introduced in 0.23.0 release.
Do not allow to explicitely specify a new token owner. Rely on the
`AnySigner` result.
The above was an improvement. We deprecate `x/username` pacakge and do
not want to introduce more work, even if those are fixing issues.

resolve #1066