orm: range query implementation
A new query mod `range` was implemented.  Using query data, it is
possible to declare start and end of a query. Each result is limited to
certain amount of results.
For bucket range query, data format is <start>[:<end>]
For index queries, format is  <start>[:<offset>[:<end>]]
Start is inclusive, end is exclusive. All values must be hex encoded.

resolve #1107