Termdeposit base rates (#1130)
* x/temrmdeposit: add custom qscore list to configuration

Update `x/termdeposit` extension and allow to manage a list of custom
q-score values for addresses. Managed list is not used for any
on chain computation. Values are stored for transparency and used in an
off chain reward computation process.

reward #1122

* a new `weave.Fraction` type introduced

We use custom fraction type implementation in many places. It is time to
extract this common functionality.
Not all existing fraction implementation can be replaced, because they
are not binary compatible with each other.

`qualityscore` and `termdeposit` was updated to user `weave.Fraction`
instead of a floating point type.